made all flags long; before some were incorrectly declared as int

+++ src/beagle.cpp
@@ -54,8 +54,8 @@ int createInstance(
 				int matrixBufferCount,		/**< Number of rate matrix buffers (input) */
 				int* resourceList,			/**< List of potential resource on which this instance is allowed (input, NULL implies no restriction */
 				int resourceCount,			/**< Length of resourceList list (input) */
-				int preferenceFlags,		/**< Bit-flags indicating preferred implementation charactertistics, see BeagleFlags (input) */
-				int requirementFlags		/**< Bit-flags indicating required implementation characteristics, see BeagleFlags (input) */
+				long preferenceFlags,		/**< Bit-flags indicating preferred implementation charactertistics, see BeagleFlags (input) */
+				long requirementFlags		/**< Bit-flags indicating required implementation characteristics, see BeagleFlags (input) */
 				)
 {
 	try {
